Move Modal component to components/ui and fix imports
