+++ airbyte-webapp/src/components/ConfirmationModal/ConfirmationModal.tsx
@@ -4,7 +4,7 @@ import styled from "styled-components";
 
 import { LoadingButton } from "components";
 import { Button } from "components/base/Button";
-import Modal from "components/Modal";
+import { Modal } from "components/ui/Modal";
 
 import useLoadingState from "../../hooks/useLoadingState";
