+++ airbyte-webapp/src/components/index.tsx
@@ -10,7 +10,6 @@ export * from "./LabeledSwitch";
 export * from "./Link";
 export * from "./LoadingPage";
 export * from "./MainPageWithScroll";
-export * from "./Modal";
 export * from "./PageTitle";
 export * from "./SimpleTableComponents";
 export * from "./StatusIcon";
