+++ airbyte-webapp/src/components/ui/Modal/Modal.module.scss
@@ -1,6 +1,6 @@
-@use "../../scss/colors";
-@use "../../scss/variables";
-@use "../../scss/z-indices";
+@use "../../../scss/colors";
+@use "../../../scss/variables";
+@use "../../../scss/z-indices";
 
 @keyframes fade-in-container {
   from {
