+++ airbyte-webapp/src/components/ui/Modal/ModalBody.module.scss
@@ -1,4 +1,4 @@
-@use "../../scss/variables";
+@use "../../../scss/variables";
 
 .modalBody {
   padding: variables.$spacing-lg variables.$spacing-xl;
