+++ airbyte-webapp/src/components/ui/Modal/ModalFooter.module.scss
@@ -1,5 +1,5 @@
-@use "../../scss/colors";
-@use "../../scss/variables";
+@use "../../../scss/colors";
+@use "../../../scss/variables";
 
 .modalFooter {
   display: flex;
