+++ airbyte-webapp/src/components/ui/Modal/index.ts
@@ -1,8 +1,3 @@
-import Modal from "./Modal";
-
 export * from "./Modal";
 export * from "./ModalBody";
 export * from "./ModalFooter";
-
-export { Modal };
-export default Modal;
