+++ airbyte-webapp/src/views/Connection/CatalogDiffModal/CatalogDiffModal.tsx
@@ -5,7 +5,7 @@ import { Button } from "components";
 
 import { AirbyteCatalog, CatalogDiff } from "core/request/AirbyteClient";
 
-import { ModalBody, ModalFooter } from "../../../components/Modal";
+import { ModalBody, ModalFooter } from "../../../components/ui/Modal";
 import styles from "./CatalogDiffModal.module.scss";
 import { DiffSection } from "./components/DiffSection";
 import { FieldSection } from "./components/FieldSection";
