+++ airbyte-webapp/src/views/Connector/ServiceForm/components/Sections/VariableInputFieldForm.tsx
@@ -4,7 +4,8 @@ import { FormattedMessage } from "react-intl";
 import { useAsync, useEffectOnce } from "react-use";
 import * as yup from "yup";
 
-import { Button, ModalBody, ModalFooter } from "components";
+import { Button } from "components/base/Button";
+import { ModalBody, ModalFooter } from "components/ui/Modal";
 
 import { FormGroupItem, FormObjectArrayItem } from "core/form/types";
